refactor(compare): hoist compare-URL sync into a headless page-level component
useCompareUrlSync() lived in CompareTable, coupling the compare-id → URL
projection to one renderer. Move it into a dedicated headless <CompareUrlSync/>
rendered by the compare page, so the projection survives swapping the table and
CompareTable stays purely presentational. Page-scoped on purpose: CompareProvider
is global, so syncing here (not in the provider) avoids writing ?ids= on other routes.

The projection was a single-use hook (useCompareUrlSync) with only this one caller,
so it's inlined into the component and the hook file removed.
